round sentiment score down

+++ portfolio/src/main/java/com/google/sps/servlets/DataServlet.java
@@ -21,6 +21,9 @@ import com.google.appengine.api.datastore.PreparedQuery;
 import com.google.appengine.api.datastore.Query;
 import com.google.appengine.api.datastore.Query.SortDirection;
 import com.google.appengine.api.datastore.FetchOptions;
+import com.google.cloud.language.v1.Document;
+import com.google.cloud.language.v1.LanguageServiceClient;
+import com.google.cloud.language.v1.Sentiment;
 import java.io.IOException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -48,7 +51,7 @@ public class DataServlet extends HttpServlet {
         Document.newBuilder().setContent(body).setType(Document.Type.PLAIN_TEXT).build();
       LanguageServiceClient languageService = LanguageServiceClient.create();
       Sentiment sentiment = languageService.analyzeSentiment(doc).getDocumentSentiment();
-      float score = sentiment.getScore();
+      double score = sentiment.getScore();
       languageService.close();
 
 
@@ -76,8 +79,8 @@ public class DataServlet extends HttpServlet {
       String title;
       String body;
       String email;
-      float score;
-      public Comment(String title, String body,String email, float score){
+      double score;
+      public Comment(String title, String body,String email, double score){
           this.title = title;
           this.body = body;
           this.email = email;
@@ -92,7 +95,7 @@ public class DataServlet extends HttpServlet {
       public String getEmail(){
         return email;
       }
-      public float getScore(){
+      public double getScore(){
         return score;
       }
   }
@@ -146,7 +149,7 @@ public void doGet(HttpServletRequest request, HttpServletResponse response) thro
     String title = (String) entity.getProperty("title");
     String body = (String) entity.getProperty("body");
     String email = (String) entity.getProperty("email");
-    float score = (float) entity.getProperty("score");
+    double score = (double) entity.getProperty("score");
     Comment comment = new Comment(title, body, email, score);
     return comment;
 }
